show parameter of methods
